RUMM-837 Move 'privacy' package into the right parent
